backend/scripts/python/requirements.txt
nfl-data-py
psycopg2-binary
pandas
python-dotenv